reorganize plugin folders
