+++ src/plugins/AnnotationEditor/components/InfomationPanel.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import dynamic from "next/dynamic";
 import { type PluginInformationPanel } from "src/index";
 import styles from "./InformationPanel.module.css";
-import { useEditorState } from "./annotation-editor-context";
+import { useEditorState } from "../context/annotation-editor-context";
 
 const AnnotationItem = dynamic(() => import("./AnnotationItem"), {
   ssr: false,
