+++ src/plugins/AnnotationEditor/components/InformationPanel.module.css
