+++ src/plugins/AnnotationEditor/utils/annotation-utils.ts
